Atlas: [Splitting The Blob: Conversion]
**Tangle:** `src/semantic/conversion.rs` was a monolithic file (~3000 lines) mixing statement classification with value extraction.
**Blueprint:** Split `src/semantic/conversion.rs` into `classification.rs` and `extraction.rs` within a new `src/semantic/conversion/` module. Added `mod.rs` as a facade to keep the internal API surface unchanged. Tests were cleanly separated into `tests.rs` to avoid circular dependencies.
