+++ .jules/atlas.md
@@ -20,3 +20,7 @@
 **Tangle:** Several modules under `src/tools/` (specifically `cache`, `report`, and `ui`) and `src/semantic/assembly/` (`model`) were exposed as `pub mod`, breaking encapsulation by exposing internal implementation details to the public API.
 **Blueprint:** Modified `src/tools/mod.rs` and `src/semantic/assembly/mod.rs` to restrict these modules with `pub(crate) mod`.
 **Stability:** Achieved higher cohesion by keeping the public API surface minimal and ensuring internal structures don't leak out of their intended domains.
+
+## [Splitting The Blob: Conversion]
+**Tangle:** `src/semantic/conversion.rs` was a monolithic file (~3000 lines) mixing statement classification with value extraction.
+**Blueprint:** Split `src/semantic/conversion.rs` into `classification.rs` and `extraction.rs` within a new `src/semantic/conversion/` module. Added `mod.rs` as a facade to keep the internal API surface unchanged. Tests were cleanly separated into `tests.rs` to avoid circular dependencies.
